[OverlayX] Move overlay types to own headers
