+++ src/client/client_engine.cpp
@@ -7,7 +7,8 @@
 #include "gl/primitive.h"
 #include "gl/vertex_array.h"
 #include "gui/gui_constants.h"
-#include "gui/overlay.h"
+#include "gui/overlay_factory.h"
+#include "gui/overlay_stack.h"
 #include "gui/widget/label_widget.h"
 #include "input/input_state.h"
 #include "lua/client_lua_api.h"
