+++ src/client/gui/overlay_stack.cpp
@@ -1,4 +1,4 @@
-#include "overlay.h"
+#include "overlay_stack.h"
 
 namespace gui {
